--- frontend/components/dashboard/iexam.ts
@@ -1,5 +1,0 @@
-export interface iexam {
-    id: number,
-    title: string,
-    body: string,
-  }
