+++ frontend/package.json
@@ -18,7 +18,6 @@
     "@lottiefiles/lottie-player": "^1.5.6",
     "@mui/icons-material": "^5.5.1",
     "@mui/material": "^5.5.1",
-    "babel-plugin-styled-components": "^2.0.6",
     "next": "12.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
